Move jwt logics to separate packages
