+++ src/main/scala/modules/jwt/validator/JwtValidator.scala
@@ -1,4 +1,4 @@
-package modules.jwt
+package modules.jwt.validator
 
 import scala.util.Try
